+++ examples/python_methods.py
@@ -1,14 +1,5 @@
 from interpreter.basic_value import BasicValue
 
-# pyimport("examples/python_methods.py", "__testq__", "dictionaries");
-# pyimport("examples/python_methods.py", "__testw__", "lists");
-
-def dictionaries(arguments):
-    return BasicValue({"A": "B"})
-
-def lists(arguments):
-    return BasicValue(["A", "B"])
-
 def p(arguments):
     if len(arguments.arguments) > 0:
         print(arguments.arguments[0].extract_value())
@@ -17,7 +8,9 @@ def p(arguments):
     return BasicValue("Nothing provided.")
 
 # Now just run:
-# pyimport("examples/python_methods.py", "__p__", "p");
-# pyimport("examples/python_methods.py", ["__p__"], ["p"]);
-# pyimport("examples/python_methods.py", Dict.new(["__p__"], ["p"]));
-# To add it to ParaCode
+# pyimport("examples/python_methods.py", ["p": "__p__"]);
+# To add it to ParaCode.
+# The first parameter is the python file's path
+# The second parameter is a dictionary where the keys
+# are the python function names and the values are the
+# ParaCode function names.