+++ CHANGELOG.md
@@ -22,6 +22,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 - Made the repl slightly nicer to look at
+- Made dictionaries more internal.
 
 ## [2.0.1] - 2021-10-18
 ### Fixed